Move retrieval code into two separate Strategy classes
The method `getLatestAPKDetails()` retrieves download information for .apk-files of the specific app source, which is in this case GitLab.com. There are currently two retrieval strategies implemented in this method: Gitlab API and Gitlab's tags-page. Both are Gitlab specific strategies used in a certain logic (API-token available or not). This refactoring step should decouple this retrieval code.